chore: rename code-gen to codegen
