+++ lib/prepare/index.js
@@ -1,7 +1,7 @@
 const path = require('path')
 const fs = require('fs-extra')
 const resolveOptions = require('./resolveOptions')
-const { genRoutesFile, genComponentRegistrationFile } = require('./code-gen')
+const { genRoutesFile, genComponentRegistrationFile } = require('./codegen')
 const { writeTemp, writeEnhanceTemp } = require('./util')
 
 module.exports = async function prepare (sourceDir) {
